requirements.txt
python-dotenv~=1.0.1

sqlalchemy~=2.0.38
pymysql~=1.1.1

openai~=1.61.1
python-telegram-bot[job-queue]~=21.10
